Add scipt
This repository serves the purpose of creating a simple scraper which
can convert html pages on online-literature.com to markdown files. As
it currently stands, the scraper takes a page link as input and creates
a markdown file for every chapter found on the provided page.

Options to change the output directory are also available.